WORKAROUND: disable some advanced features depending on newer kernel
We want to demo hotspot on a machine that runs 16.04 and apparently
doesn't work with newer kernels. That kernel does not support the
perf record features for --sample-cpu and --switch-events, so disable
these features for now. Also, disable the elevate privileges
feature, as it does not work correctly from within an AppImage.